# Optimize `Value::unit()` to avoid additional allocations